apple-music-feed/requirements.txt
PyJWT
requests
cryptography
pandas
pyarrow